requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.0.1
langchain-groq>=0.0.1
supabase>=1.0.0
python-dotenv>=1.0.0
cohere>=4.0.0
groq>=0.1.0
neo4j>=5.0.0
